The login section is now complete and fully functional. As a regular user, you can't get to the admin pages anymore.
Plus some additional changes like when you are logged in, at the top it now says log out instead of log in and you can log out when you click on it and the session will be destroyed.

+++ Webshop/deleteProduct.php
@@ -2,6 +2,16 @@
 require_once 'includes/database.php';
 /** @var $db */
 
+session_start();
+
+$login = $_SESSION['login'];
+$admin = $_SESSION['admin'];
+
+if($admin != 1 || $login == false) {
+    header("Location: login.php");
+    exit();
+}
+
 $productId = $_GET['id'];
 
 $query = "DELETE FROM garen WHERE id = '$productId'";
